Min/max support for dates; numeric input; split form util files
