+++ src/components/elements/input/DatePicker.tsx
@@ -4,10 +4,13 @@ import {
   DatePicker as MuiDatePicker,
   DatePickerProps,
 } from '@mui/x-date-pickers';
+import { isPast } from 'date-fns';
+import { useMemo } from 'react';
 
 import TextInput, { TextInputProps } from './TextInput';
 
-import { DynamicInputCommonProps } from '@/modules/form/formUtil';
+import { DynamicInputCommonProps } from '@/modules/form/components/DynamicField';
+import { isDate } from '@/modules/form/util/formUtil';
 
 interface PickerProps
   extends Omit<DatePickerProps<DateType, DateType>, 'renderInput'> {
@@ -17,18 +20,32 @@ interface PickerProps
 
 type Props = PickerProps & DynamicInputCommonProps;
 
-const DatePicker: React.FC<Props> = ({
+const DatePicker = ({
   sx,
   textInputProps,
   error,
+  min,
+  max,
   ...props
-}: Props) => (
-  <MuiDatePicker
-    {...props}
-    renderInput={(params) => (
-      <TextInput sx={sx} {...textInputProps} {...params} error={error} />
-    )}
-  />
-);
+}: Props) => {
+  const defaultOpenMonth = useMemo(() => {
+    if (isDate(min)) return min;
+    if (isDate(max) && isPast(max)) return max;
+    return undefined;
+  }, [min, max]);
+
+  return (
+    <MuiDatePicker
+      {...props}
+      showDaysOutsideCurrentMonth
+      minDate={isDate(min) ? min : undefined}
+      maxDate={isDate(max) ? max : undefined}
+      defaultCalendarMonth={defaultOpenMonth}
+      renderInput={(params) => (
+        <TextInput sx={sx} {...textInputProps} {...params} error={error} />
+      )}
+    />
+  );
+};
 
 export default DatePicker;
