+++ src/modules/form/util/recordFormUtil.ts
@@ -1,0 +1,207 @@
+/**
+ * Utils specifically for the EditRecord component
+ */
+
+import { compact, isNil } from 'lodash-es';
+
+import { formatDateForGql, parseHmisDateString } from '../../hmis/hmisUtil';
+
+import {
+  getOptionValue,
+  isDataNotCollected,
+  resolveOptionList,
+} from './formUtil';
+
+import {
+  FormDefinitionJson,
+  FormItem,
+  ItemType,
+  PickListOption,
+} from '@/types/gqlTypes';
+
+/**
+ * Transform form value shape into GraphQL value shape.
+ * For example, a selected option '{ code: 'ES' }'
+ * is converted to enum 'ES'.
+ *
+ * @param value value from the form state
+ * @param item corresponding FormDefinition item
+ * @returns transformed value
+ */
+export const formValueToGqlValue = (value: any, item: FormItem): any => {
+  if (value instanceof Date) {
+    return formatDateForGql(value);
+  }
+
+  if (item.type === ItemType.Integer) {
+    return parseInt(value);
+  }
+
+  if (
+    item.type === ItemType.Choice &&
+    item.pickListReference &&
+    ['projects', 'organizations'].includes(item.pickListReference)
+  ) {
+    // Special case for project/org selectors which key on ID
+    if (Array.isArray(value)) {
+      return value.map((option: { id: string }) => option.id);
+    } else if (value) {
+      return (value as { id: string }).id;
+    }
+  } else if ([ItemType.Choice, ItemType.OpenChoice].includes(item.type)) {
+    if (Array.isArray(value)) {
+      return value.map((option: PickListOption) => option.code);
+    } else if (value) {
+      return (value as PickListOption).code;
+    }
+  }
+
+  if (value === '') {
+    return undefined;
+  }
+  return value;
+};
+
+const findDataNotCollectedCode = (item: FormItem): string | undefined => {
+  const options = resolveOptionList(item, true) || [];
+
+  return options.find(
+    (opt) => isDataNotCollected(opt.code) || opt.code === 'NOT_APPLICABLE'
+  )?.code;
+};
+
+type TransformSubmitValuesParams = {
+  definition: FormDefinitionJson;
+  /** form state (from DynamicForm) to transform */
+  values: Record<string, any>;
+  /** whether to fill unanswered questions with Data Not Collected option (if present) */
+  autofillNotCollected?: boolean;
+  /** whether to fill unanswered questions with `null` */
+  autofillNulls?: boolean;
+  /** whether to fill unanswered boolean questions `false` */
+  autofillBooleans?: boolean;
+};
+
+/**
+ * Given the form state of a completed form, transform it into
+ * query variables for a mutation. This is used for dynamic forms that
+ * edit one record directly, like the Client, Project, and Organization edit screens.
+ */
+export const transformSubmitValues = ({
+  definition,
+  values,
+
+  autofillNotCollected = false,
+  autofillNulls = false,
+  autofillBooleans = false,
+}: TransformSubmitValuesParams) => {
+  const result: Record<string, any> = {};
+
+  // Recursive helper for traversing the FormDefinition
+  function recursiveTransformValues(
+    items: FormItem[],
+    values: Record<string, any>,
+    transformed: Record<string, any> // result map to be filled in
+  ) {
+    items.forEach((item: FormItem) => {
+      if (Array.isArray(item.item)) {
+        recursiveTransformValues(item.item, values, transformed);
+      }
+
+      const key = item.queryField;
+      if (!key) return;
+
+      let value;
+      if (item.linkId in values) {
+        // Transform into gql value, for example Date -> YYYY-MM-DD string
+        value = formValueToGqlValue(values[item.linkId], item);
+      }
+
+      if (typeof value !== 'undefined') {
+        transformed[key] = value;
+      }
+
+      if (autofillNotCollected && isNil(value)) {
+        // If we don't have a value, fill in Not Collected code if present
+        const notCollectedCode = findDataNotCollectedCode(item);
+        if (notCollectedCode) transformed[key] = notCollectedCode;
+      }
+      if (autofillNulls && isNil(transformed[key])) {
+        transformed[key] = null;
+      }
+      if (
+        autofillBooleans &&
+        isNil(transformed[key]) &&
+        item.type === ItemType.Boolean
+      ) {
+        transformed[key] = false;
+      }
+    });
+  }
+
+  recursiveTransformValues(definition.item || [], values, result);
+
+  return result;
+};
+
+// Convert gql value into value for form state
+const getFormValue = (value: any | null | undefined, item: FormItem) => {
+  if (isNil(value)) return value;
+
+  switch (item.type) {
+    case ItemType.Date:
+      return parseHmisDateString(value);
+
+    case ItemType.Choice:
+    case ItemType.OpenChoice:
+      if (Array.isArray(value)) {
+        // DNC gets filtered out here
+        return compact(value.map((v) => getOptionValue(v, item)));
+      }
+      return getOptionValue(value, item);
+
+    default:
+      // Set the property directly as the initial form value
+      return value;
+  }
+};
+
+/**
+ * Create initial form values based on a record.
+ * This is only used for forms that edit a record directly, like the Client, Project, and Organization edit screens.
+ *
+ * @param definition FormDefinition
+ * @param record  GQL HMIS record, like Project or Organization
+ *
+ * @returns initial form state, ready to pass to DynamicForm as initialValues
+ */
+export const createInitialValuesFromRecord = (
+  definition: FormDefinitionJson,
+  record: any
+): Record<string, any> => {
+  const initialValues: Record<string, any> = {};
+
+  // Recursive helper for traversing the FormDefinition
+  function recursiveFillInValues(
+    items: FormItem[],
+    record: Record<string, any>,
+    values: Record<string, any> // intialValues object to be filled in
+  ) {
+    items.forEach((item: FormItem) => {
+      if (Array.isArray(item.item)) {
+        recursiveFillInValues(item.item, record, values);
+      }
+      // Skip: this question doesn't map to a field
+      if (!item.queryField) return;
+
+      // Skip: the record doesn't have a value for this property
+      if (!record.hasOwnProperty(item.queryField)) return;
+
+      values[item.linkId] = getFormValue(record[item.queryField], item);
+    });
+  }
+
+  recursiveFillInValues(definition.item || [], record, initialValues);
+
+  return initialValues;
+};
